# ALUMNOS/MDAB/FRANCISCO_ALVAREZ_V/API/exercises/openapi-exercises/02-postman-integration/books-api-openapi.yaml
openapi: 3.0.3
info:
  title: Books API
  description: A well-documented API for managing books
  version: 1.0.0

servers:
  - url: http://127.0.0.1:5000
    description: Local development server

tags:
  - name: books
    description: Book management operations

paths:
  /api/books:
    get:
      tags:
        - books
      summary: List all books
      description: Returns a list of books with optional filtering and pagination
      operationId: listBooks
      parameters:
        - name: author
          in: query
          description: Filter books by author name (partial match)
          required: false
          schema:
            type: string
            example: Orwell
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of books per page
          required: false
          schema:
            type: integer
            default: 10
            example: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
              example:
                - id: 1
                  title: "1984"
                  author: "George Orwell"
                  year: 1949
                  isbn: "978-0451524935"
                - id: 2
                  title: "To Kill a Mockingbird"
                  author: "Harper Lee"
                  year: 1960
                  isbn: "978-0061120084"

    post:
      tags:
        - books
      summary: Create a new book
      description: Add a new book to the collection
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
            example:
              title: "1984"
              author: "George Orwell"
              year: 1949
              isbn: "978-0451524935"
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/books/{id}:
    get:
      tags:
        - books
      summary: Get a book by ID
      description: Returns a single book if it exists
      operationId: getBook
      parameters:
        - name: id
          in: path
          description: The book identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - books
      summary: Update a book
      description: Update book information (all fields optional)
      operationId: updateBook
      parameters:
        - name: id
          in: path
          description: The book identifier
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInput'
            example:
              title: "1984 (Updated)"
              author: "George Orwell"
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - books
      summary: Delete a book
      description: Permanently remove a book from the system
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          description: The book identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - author
      properties:
        id:
          type: integer
          description: The book unique identifier
          example: 1
        title:
          type: string
          description: Book title
          example: "1984"
        author:
          type: string
          description: Book author
          example: "George Orwell"
        year:
          type: integer
          description: Publication year
          example: 1949
        isbn:
          type: string
          description: ISBN number
          example: "978-0451524935"

    BookInput:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          description: Book title
          example: "1984"
        author:
          type: string
          description: Book author
          example: "George Orwell"
        year:
          type: integer
          description: Publication year
          example: 1949
        isbn:
          type: string
          description: ISBN number
          example: "978-0451524935"

    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Book not found"